Aggregation/requirements.txt
aiokafka==0.10.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncio==3.4.3
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
packaging==23.2
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
